+++ src/Platform/Apple/OSXInput.mm
@@ -235,7 +235,7 @@ namespace
         m_appActive = activate;
     }
 
-    void Input::WindowFeedMouseWheel(int32 delta, Window* window)
+    void Input::WindowFeedMouseWheel(float delta, Window* window)
     {
         m_mouseScroll = delta;
